requirements.txt
fastapi
uvicorn[standard]
python-multipart
pandas
openpyxl
python-levenshtein
nanoid
python-dotenv
sqlalchemy 
psycopg2-binary